transition iceberg filtering to use built in functions (#21)
The iceberg filtering was originally done by iterating over all of the icebergs, performing a series of checks on each one. This PR transitions that workflow to utilizing more of the geopandas and xarray built-in functions, along with geocube and groupby, to do more of those checks "en mass", while still only ever extracting data values from the DEMs for the most promising icebergs/tests that need them.